Major changes to ToolsPanel
Added sub-components and wired up form and function. There is still a
bug where if you change the form/function after selecting the region to
mask, you'll get an error about coords not being set. I believe this is
due to all Context variables getting reloaded whenever a set method is
called, causing a redraw (and clearing of local variables). Need to
think of a way to handle this.